Split regression tests into separate file
